+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-16T08:09:01.077Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-16T12:54:27.109Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")